use float[][] for score matrix

+++ src/main/java/edu/neu/ccs/pyramid/classification/lkboost/LKBoostOptimizer.java
@@ -106,7 +106,7 @@ public class LKBoostOptimizer extends GBOptimizer {
      */
     private void updateClassProb(int i){
         int numClasses = this.boosting.getNumClasses();
-        double[] scores = scoreMatrix.getScoresForData(i);
+        float[] scores = scoreMatrix.getScoresForData(i);
 
         double logDenominator = MathUtil.logSumExp(scores);
 //        if (logger.isDebugEnabled()){
